- fix sql error in test - fix empty string for title
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@14758 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tests/tine20/Admin/JsonTest.php
@@ -402,7 +402,9 @@ class Admin_JsonTest extends PHPUnit_Framework_TestCase
         }
         
         // delete logs
-        $this->_backend->deleteAccessLogEntries($deleteLogIds);
+        if (!empty($deleteLogIds)) {
+            $this->_backend->deleteAccessLogEntries($deleteLogIds);
+        }
         
         // check total count
         $accessLogs = $this->_backend->getAccessLogEntries($from->get(Tinebase_Record_Abstract::ISO8601LONG), $to->get(Tinebase_Record_Abstract::ISO8601LONG), 'tine20admin', $paging);
